- integration QBChatDialog (in progress)
