+++ Q-municate_app/src/main/java/com/quickblox/q_municate/ui/activities/chats/AddFriendsToGroupActivity.java
@@ -91,7 +91,7 @@ public class AddFriendsToGroupActivity extends BaseFriendsListActivity {
     protected void performDone() {
         List<User> selectedFriendsList = ((SelectableFriendsAdapter) friendsAdapter).getSelectedFriendsList();
         if (!selectedFriendsList.isEmpty()) {
-            boolean joined = groupChatHelper != null && groupChatHelper.isDialogJoined(qbDialog);
+            boolean joined = chatHelper != null && chatHelper.isDialogJoined(qbDialog);
             if (isChatInitializedAndUserLoggedIn() && checkNetworkAvailableWithError() && joined) {
                 showProgress();
                 friendIdsList = UserFriendUtils.getFriendIds(selectedFriendsList);
