+++ Q-municate_app/src/main/java/com/quickblox/q_municate/ui/activities/profile/UserProfileActivity.java
@@ -29,7 +29,6 @@ import com.quickblox.q_municate_core.utils.OnlineStatusUtils;
 import com.quickblox.q_municate_core.utils.UserFriendUtils;
 import com.quickblox.q_municate_db.managers.DataManager;
 import com.quickblox.q_municate_db.managers.UserDataManager;
-import com.quickblox.q_municate_db.models.Dialog;
 import com.quickblox.q_municate_db.models.DialogOccupant;
 import com.quickblox.q_municate_db.models.User;
 import com.quickblox.users.model.QBUser;
@@ -279,7 +278,7 @@ public class UserProfileActivity extends BaseLoggableActivity {
             finish();
         } else {
             String dialogId = dialogOccupant.getDialog().getDialogId();
-            QBDeleteChatCommand.start(this, dialogId, Dialog.Type.PRIVATE);
+            QBDeleteChatCommand.start(this, dialogId);
         }
     }
