+++ Q-municate_core/src/main/java/com/quickblox/q_municate_core/qb/commands/QBLoadAttachFileCommand.java
@@ -6,7 +6,7 @@ import android.os.Bundle;
 
 import com.quickblox.content.model.QBFile;
 import com.quickblox.q_municate_core.core.command.ServiceCommand;
-import com.quickblox.q_municate_core.qb.helpers.QBPrivateChatHelper;
+import com.quickblox.q_municate_core.qb.helpers.QBChatHelper;
 import com.quickblox.q_municate_core.service.QBService;
 import com.quickblox.q_municate_core.service.QBServiceConsts;
 
@@ -16,12 +16,12 @@ public class QBLoadAttachFileCommand extends ServiceCommand {
 
     private static final String TAG = QBLoadAttachFileCommand.class.getSimpleName();
 
-    private final QBPrivateChatHelper privateChatHelper;
+    private final QBChatHelper chatHelper;
 
-    public QBLoadAttachFileCommand(Context context, QBPrivateChatHelper privateChatHelper,
+    public QBLoadAttachFileCommand(Context context, QBChatHelper chatHelper,
             String successAction, String failAction) {
         super(context, successAction, failAction);
-        this.privateChatHelper = privateChatHelper;
+        this.chatHelper = chatHelper;
     }
 
     public static void start(Context context, File file, String dialogId) {
@@ -36,7 +36,7 @@ public class QBLoadAttachFileCommand extends ServiceCommand {
         File file = (File) extras.getSerializable(QBServiceConsts.EXTRA_FILE);
         String dialogId = (String) extras.getSerializable(QBServiceConsts.EXTRA_DIALOG_ID);
 
-        QBFile qbFile = privateChatHelper.loadAttachFile(file);
+        QBFile qbFile = chatHelper.loadAttachFile(file);
 
         Bundle result = new Bundle();
         result.putSerializable(QBServiceConsts.EXTRA_ATTACH_FILE, qbFile);
