+++ Q-municate_core/src/main/java/com/quickblox/q_municate_core/qb/commands/chat/QBJoinGroupChatsCommand.java
@@ -5,18 +5,18 @@ import android.content.Intent;
 import android.os.Bundle;
 
 import com.quickblox.q_municate_core.core.command.ServiceCommand;
-import com.quickblox.q_municate_core.qb.helpers.QBGroupChatHelper;
+import com.quickblox.q_municate_core.qb.helpers.QBChatHelper;
 import com.quickblox.q_municate_core.service.QBService;
 import com.quickblox.q_municate_core.service.QBServiceConsts;
 
 public class QBJoinGroupChatsCommand extends ServiceCommand {
 
-    private QBGroupChatHelper multiChatHelper;
+    private QBChatHelper chatHelper;
 
-    public QBJoinGroupChatsCommand(Context context, QBGroupChatHelper multiChatHelper, String successAction,
-            String failAction) {
+    public QBJoinGroupChatsCommand(Context context, QBChatHelper chatHelper, String successAction,
+                                   String failAction) {
         super(context, successAction, failAction);
-        this.multiChatHelper = multiChatHelper;
+        this.chatHelper = chatHelper;
     }
 
     public static void start(Context context) {
@@ -26,7 +26,7 @@ public class QBJoinGroupChatsCommand extends ServiceCommand {
 
     @Override
     public Bundle perform(Bundle extras) throws Exception {
-        multiChatHelper.tryJoinRoomChats();
+        chatHelper.tryJoinRoomChats();
         return null;
     }
 }
