+++ Q-municate_core/src/main/java/com/quickblox/q_municate_core/qb/commands/friend/QBInitFriendListCommand.java
@@ -5,8 +5,8 @@ import android.content.Intent;
 import android.os.Bundle;
 
 import com.quickblox.q_municate_core.core.command.ServiceCommand;
+import com.quickblox.q_municate_core.qb.helpers.QBChatHelper;
 import com.quickblox.q_municate_core.qb.helpers.QBFriendListHelper;
-import com.quickblox.q_municate_core.qb.helpers.QBPrivateChatHelper;
 import com.quickblox.q_municate_core.service.QBService;
 import com.quickblox.q_municate_core.service.QBServiceConsts;
 
@@ -15,13 +15,13 @@ public class QBInitFriendListCommand extends ServiceCommand {
     private final static String TAG = QBInitFriendListCommand.class.getSimpleName();
 
     private QBFriendListHelper friendListHelper;
-    private QBPrivateChatHelper privateChatHelper;
+    private QBChatHelper chatHelper;
 
-    public QBInitFriendListCommand(Context context, QBFriendListHelper friendListHelper, QBPrivateChatHelper privateChatHelper, String successAction,
-            String failAction) {
+    public QBInitFriendListCommand(Context context, QBFriendListHelper friendListHelper, QBChatHelper chatHelper, String successAction,
+                                   String failAction) {
         super(context, successAction, failAction);
         this.friendListHelper = friendListHelper;
-        this.privateChatHelper = privateChatHelper;
+        this.chatHelper = chatHelper;
     }
 
     public static void start(Context context) {
@@ -31,7 +31,7 @@ public class QBInitFriendListCommand extends ServiceCommand {
 
     @Override
     public Bundle perform(Bundle extras) throws Exception {
-        friendListHelper.init(privateChatHelper);
+        friendListHelper.init(chatHelper);
         return extras;
     }
 }
